+++ sympy/polys/rootisolation.py
@@ -1,26 +1,602 @@
 """Tools for real and complex root isolation and refinement. """
 
 from sympy.polys.densebasic import (
-    dup_LC, dup_degree,
+    dup_LC, dup_TC, dup_degree,
+    dup_strip, dup_reverse,
     dup_convert, dmp_convert,
     dup_terms_gcd,
 )
 
+from sympy.polys.densearith import (
+    dup_neg, dup_rshift,
+)
+
 from sympy.polys.densetools import (
-    dup_isolate_real_roots_list,
-    dup_isolate_real_roots_sqf,
-    dup_refine_real_root,
+    dup_scale, dup_taylor, dup_mirror,
+    dup_transform,
     dup_eval, dmp_eval_in,
+    dup_sign_variations,
     dup_clear_denoms,
     dup_real_imag,
     dup_sqf_list,
 )
 
+from sympy.polys.factortools import (
+    dup_factor_list,
+)
+
 from sympy.polys.polyerrors import (
+    RefinementFailed,
     DomainError,
 )
 
 import operator
+
+def dup_root_upper_bound(f, K):
+    """Compute LMQ upper bound for `f`'s positive roots. """
+    n, t, P = len(f), K.one, []
+
+    if dup_LC(f, K) < 0:
+        f = dup_neg(f, K)
+
+    f = list(reversed(f))
+
+    for i in xrange(0, n):
+        if f[i] >= 0:
+            continue
+
+        a, Q = K.log(-f[i], 2), []
+
+        for j in xrange(i+1, n):
+
+            if f[j] <= 0:
+                continue
+
+            q = t + a - K.log(f[j], 2)
+            Q.append(q // (j - i))
+
+            t += 1
+
+        if not Q:
+            continue
+
+        P.append(min(Q))
+
+    if not P:
+        return None
+    else:
+        return 2.0**(max(P)+1)
+
+def dup_root_lower_bound(f, K):
+    """Compute LMQ lower bound for `f`'s positive roots. """
+    bound = dup_root_upper_bound(dup_reverse(f), K)
+
+    if bound is not None:
+        return 1.0 / bound
+    else:
+        return None
+
+def _mobius_from_interval(I, field):
+    """Convert an open interval to a Mobius transform. """
+    s, t = I
+
+    a, c = field.numer(s), field.denom(s)
+    b, d = field.numer(t), field.denom(t)
+
+    return a, b, c, d
+
+def _mobius_to_interval(M, field):
+    """Convert a Mobius transform to an open interval. """
+    a, b, c, d = M
+
+    s, t = field(a, c), field(b, d)
+
+    if s <= t:
+        return (s, t)
+    else:
+        return (t, s)
+
+def dup_step_refine_real_root(f, M, K, fast=False):
+    """One step of positive real root refinement algorithm. """
+    a, b, c, d = M
+
+    if a == b and c == d:
+        return f, (a, b, c, d)
+
+    A = dup_root_lower_bound(f, K)
+
+    if A is not None:
+        A = K(int(A))
+    else:
+        A = K.zero
+
+    if fast and A > 16:
+        f = dup_scale(f, A, K)
+        a, c, A = A*a, A*c, K.one
+
+    if A >= K.one:
+        f = dup_taylor(f, A, K)
+        b, d = A*a + b, A*c + d
+
+        if not dup_eval(f, K.zero, K):
+            return f, (b, b, d, d)
+
+    f, g = dup_taylor(f, K.one, K), f
+
+    a1, b1, c1, d1 = a, a+b, c, c+d
+
+    if not dup_eval(f, K.zero, K):
+        return f, (b1, b1, d1, d1)
+
+    k = dup_sign_variations(f, K)
+
+    if k == 1:
+        a, b, c, d = a1, b1, c1, d1
+    else:
+        f = dup_taylor(dup_reverse(g), K.one, K)
+
+        if not dup_eval(f, K.zero, K):
+            f = dup_rshift(f, 1, K)
+
+        a, b, c, d = b, a+b, d, c+d
+
+    return f, (a, b, c, d)
+
+def dup_inner_refine_real_root(f, M, K, eps=None, steps=None, disjoint=None, fast=False, mobius=False):
+    """Refine a positive root of `f` given a Mobius transform or an interval. """
+    F = K.get_field()
+
+    if len(M) == 2:
+        a, b, c, d = _mobius_from_interval(M, F)
+    else:
+        a, b, c, d = M
+
+    while not c:
+        f, (a, b, c, d) = dup_step_refine_real_root(f, (a, b, c, d), K, fast=fast)
+
+    if eps is not None and steps is not None:
+        for i in xrange(0, steps):
+            if abs(F(a, c) - F(b, d)) >= eps:
+                f, (a, b, c, d) = dup_step_refine_real_root(f, (a, b, c, d), K, fast=fast)
+            else:
+                break
+    else:
+        if eps is not None:
+            while abs(F(a, c) - F(b, d)) >= eps:
+                f, (a, b, c, d) = dup_step_refine_real_root(f, (a, b, c, d), K, fast=fast)
+
+        if steps is not None:
+            for i in xrange(0, steps):
+                f, (a, b, c, d) = dup_step_refine_real_root(f, (a, b, c, d), K, fast=fast)
+
+    if disjoint is not None:
+        while True:
+            u, v = _mobius_to_interval((a, b, c, d), F)
+
+            if v <= disjoint or disjoint <= u:
+                break
+            else:
+                f, (a, b, c, d) = dup_step_refine_real_root(f, (a, b, c, d), K, fast=fast)
+
+    if not mobius:
+        return _mobius_to_interval((a, b, c, d), F)
+    else:
+        return f, (a, b, c, d)
+
+def dup_outer_refine_real_root(f, s, t, K, eps=None, steps=None, disjoint=None, fast=False):
+    """Refine a positive root of `f` given an interval `(s, t)`. """
+    a, b, c, d = _mobius_from_interval((s, t), K.get_field())
+
+    f = dup_transform(f, dup_strip([a, b]),
+                         dup_strip([c, d]), K)
+
+    if dup_sign_variations(f, K) != 1:
+        raise RefinementFailed("there should be exactly one root in (%s, %s) interval" % (s, t))
+
+    return dup_inner_refine_real_root(f, (a, b, c, d), K, eps=eps, steps=steps, disjoint=disjoint, fast=fast)
+
+def dup_refine_real_root(f, s, t, K, eps=None, steps=None, disjoint=None, fast=False):
+    """Refine real root's approximating interval to the given precision. """
+    if K.is_QQ:
+        (_, f), K = dup_clear_denoms(f, K, convert=True), K.get_ring()
+    elif not K.is_ZZ:
+        raise DomainError("real root refinement not supported over %s" % K)
+
+    if s == t:
+        return (s, t)
+
+    if s > t:
+        s, t = t, s
+
+    negative = False
+
+    if s < 0:
+        if t <= 0:
+            f, s, t, negative = dup_mirror(f, K), -t, -s, True
+        else:
+            raise ValueError("can't refine a real root in (%s, %s)" % (s, t))
+
+    if negative and disjoint is not None:
+        if disjoint < 0:
+            disjoint = -disjoint
+        else:
+            disjoint = None
+
+    s, t = dup_outer_refine_real_root(f, s, t, K, eps=eps, steps=steps, disjoint=disjoint, fast=fast)
+
+    if negative:
+        return (-t, -s)
+    else:
+        return ( s,  t)
+
+def dup_inner_isolate_real_roots(f, K, eps=None, fast=False):
+    """Internal function for isolation positive roots up to given precision. """
+    a, b, c, d = K.one, K.zero, K.zero, K.one
+
+    k = dup_sign_variations(f, K)
+
+    if k == 0:
+        return []
+    if k == 1:
+        roots = [dup_inner_refine_real_root(f, (a, b, c, d), K, eps=eps, fast=fast, mobius=True)]
+    else:
+        roots, stack = [], [(a, b, c, d, f, k)]
+
+        F = K.get_field()
+
+        while stack:
+            a, b, c, d, f, k = stack.pop()
+
+            A = dup_root_lower_bound(f, K)
+
+            if A is not None:
+                A = K(int(A))
+            else:
+                A = K.zero
+
+            if fast and A > 16:
+                f = dup_scale(f, A, K)
+                a, c, A = A*a, A*c, K.one
+
+            if A >= K.one:
+                f = dup_taylor(f, A, K)
+                b, d = A*a + b, A*c + d
+
+                if not dup_TC(f, K):
+                    roots.append((f, (b, b, d, d)))
+                    f = dup_rshift(f, 1, K)
+
+                k = dup_sign_variations(f, K)
+
+                if k == 0:
+                    continue
+                if k == 1:
+                    roots.append(dup_inner_refine_real_root(f, (a, b, c, d), K, eps=eps, fast=fast, mobius=True))
+                    continue
+
+            f1 = dup_taylor(f, K.one, K)
+
+            a1, b1, c1, d1, r = a, a+b, c, c+d, 0
+
+            if not dup_TC(f1, K):
+                roots.append((f1, (b1, b1, d1, d1)))
+                f1, r = dup_rshift(f1, 1, K), 1
+
+            k1 = dup_sign_variations(f1, K)
+            k2 = k - k1 - r
+
+            a2, b2, c2, d2 = b, a+b, d, c+d
+
+            if k2 > 1:
+                f2 = dup_taylor(dup_reverse(f), K.one, K)
+
+                if not dup_TC(f2, K):
+                    f2 = dup_rshift(f2, 1, K)
+
+                k2 = dup_sign_variations(f2, K)
+            else:
+                f2 = None
+
+            if k1 < k2:
+                a1, a2, b1, b2 = a2, a1, b2, b1
+                c1, c2, d1, d2 = c2, c1, d2, d1
+                f1, f2, k1, k2 = f2, f1, k2, k1
+
+            if not k1:
+                continue
+
+            if f1 is None:
+                f1 = dup_taylor(dup_reverse(f), K.one, K)
+
+                if not dup_TC(f1, K):
+                    f1 = dup_rshift(f1, 1, K)
+
+            if k1 == 1:
+                roots.append(dup_inner_refine_real_root(f1, (a1, b1, c1, d1), K, eps=eps, fast=fast, mobius=True))
+            else:
+                stack.append((a1, b1, c1, d1, f1, k1))
+
+            if not k2:
+                continue
+
+            if f2 is None:
+                f2 = dup_taylor(dup_reverse(f), K.one, K)
+
+                if not dup_TC(f2, K):
+                    f2 = dup_rshift(f2, 1, K)
+
+            if k2 == 1:
+                roots.append(dup_inner_refine_real_root(f2, (a2, b2, c2, d2), K, eps=eps, fast=fast, mobius=True))
+            else:
+                stack.append((a2, b2, c2, d2, f2, k2))
+
+    return roots
+
+def _discard_if_outside_interval(f, M, inf, sup, K, negative, fast, mobius):
+    """Discard an isolating interval if outside ``(inf, sup)``. """
+    F = K.get_field()
+
+    while True:
+        u, v = _mobius_to_interval(M, F)
+
+        if negative:
+            u, v = -v, -u
+
+        if (inf is None or u >= inf) and (sup is None or v <= sup):
+            if not mobius:
+                return u, v
+            else:
+                return f, M
+        elif (sup is not None and u > sup) or (inf is not None and v < inf):
+            return None
+        else:
+            f, M = dup_step_refine_real_root(f, M, K, fast=fast)
+
+def dup_inner_isolate_positive_roots(f, K, eps=None, inf=None, sup=None, fast=False, mobius=False):
+    """Iteratively compute disjoint positive root isolation intervals. """
+    if sup is not None and sup < 0:
+        return []
+
+    roots = dup_inner_isolate_real_roots(f, K, eps=eps, fast=fast)
+
+    F, results = K.get_field(), []
+
+    if inf is not None or sup is not None:
+        for f, M in roots:
+            result = _discard_if_outside_interval(f, M, inf, sup, K, False, fast, mobius)
+
+            if result is not None:
+                results.append(result)
+    elif not mobius:
+        for f, M in roots:
+            u, v = _mobius_to_interval(M, F)
+            results.append((u, v))
+    else:
+        results = roots
+
+    return results
+
+def dup_inner_isolate_negative_roots(f, K, inf=None, sup=None, eps=None, fast=False, mobius=False):
+    """Iteratively compute disjoint negative root isolation intervals. """
+    if inf is not None and inf >= 0:
+        return []
+
+    roots = dup_inner_isolate_real_roots(dup_mirror(f, K), K, eps=eps, fast=fast)
+
+    F, results = K.get_field(), []
+
+    if inf is not None or sup is not None:
+        for f, M in roots:
+            result = _discard_if_outside_interval(f, M, inf, sup, K, True, fast, mobius)
+
+            if result is not None:
+                results.append(result)
+    elif not mobius:
+        for f, M in roots:
+            u, v = _mobius_to_interval(M, F)
+            results.append((-v, -u))
+    else:
+        results = roots
+
+    return results
+
+def _isolate_zero(f, K, inf, sup, basis=False, sqf=False):
+    """Handle special case of CF algorithm when ``f`` is homogeneous. """
+    j, f = dup_terms_gcd(f, K)
+
+    if j > 0:
+        F = K.get_field()
+
+        if (inf is None or inf <= 0) and (sup is None or 0 <= sup):
+            if not sqf:
+                if not basis:
+                    return [((F.zero, F.zero), j)], f
+                else:
+                    return [((F.zero, F.zero), j, [K.one, K.zero])], f
+            else:
+                return [(F.zero, F.zero)], f
+
+    return [], f
+
+def dup_isolate_real_roots_sqf(f, K, eps=None, inf=None, sup=None, fast=False):
+    """Isolate real roots of a square-free polynomial using CF approach. """
+    if K.is_QQ:
+        (_, f), K = dup_clear_denoms(f, K, convert=True), K.get_ring()
+    elif not K.is_ZZ:
+        raise DomainError("isolation of real roots not supported over %s" % K)
+
+    if dup_degree(f) <= 0:
+        return []
+
+    I_zero, f = _isolate_zero(f, K, inf, sup, basis=False, sqf=True)
+
+    I_neg = dup_inner_isolate_negative_roots(f, K, eps=eps, inf=inf, sup=sup, fast=fast)
+    I_pos = dup_inner_isolate_positive_roots(f, K, eps=eps, inf=inf, sup=sup, fast=fast)
+
+    return sorted(I_neg + I_zero + I_pos)
+
+def dup_isolate_real_roots(f, K, eps=None, inf=None, sup=None, basis=False, fast=False):
+    """Isolate real roots using continued fractions approach. """
+    if K.is_QQ:
+        (_, f), K = dup_clear_denoms(f, K, convert=True), K.get_ring()
+    elif not K.is_ZZ:
+        raise DomainError("isolation of real roots not supported over %s" % K)
+
+    if dup_degree(f) <= 0:
+        return []
+
+    I_zero, f = _isolate_zero(f, K, inf, sup, basis=basis, sqf=False)
+
+    _, factors = dup_sqf_list(f, K)
+
+    if len(factors) == 1:
+        ((f, k),) = factors
+
+        I_neg = dup_inner_isolate_negative_roots(f, K, eps=eps, inf=inf, sup=sup, fast=fast)
+        I_pos = dup_inner_isolate_positive_roots(f, K, eps=eps, inf=inf, sup=sup, fast=fast)
+
+        I_neg = [ ((u, v), k) for u, v in I_neg ]
+        I_pos = [ ((u, v), k) for u, v in I_pos ]
+    else:
+        I_neg, I_pos = _real_isolate_and_disjoin(factors, K,
+            eps=eps, inf=inf, sup=sup, basis=basis, fast=fast)
+
+    return sorted(I_neg + I_zero + I_pos)
+
+def dup_isolate_real_roots_list(polys, K, eps=None, inf=None, sup=None, strict=False, basis=False, fast=False):
+    """Isolate real roots of a list of square-free polynomial using CF approach. """
+    if K.is_QQ:
+        K, F, polys = K.get_ring(), K, polys[:]
+
+        for i, p in enumerate(polys):
+            polys[i] = dup_clear_denoms(p, F, K, convert=True)[1]
+    elif not K.is_ZZ:
+        raise DomainError("isolation of real roots not supported over %s" % K)
+
+    zeros, factors_dict = False, {}
+
+    if (inf is None or inf <= 0) and (sup is None or 0 <= sup):
+        zeros, zero_indices = True, {}
+
+    for i, p in enumerate(polys):
+        j, p = dup_terms_gcd(p, K)
+
+        if zeros and j > 0:
+            zero_indices[i] = j
+
+        for f, k in dup_factor_list(p, K)[1]:
+            f = tuple(f)
+
+            if f not in factors_dict:
+                factors_dict[f] = {i: k}
+            else:
+                factors_dict[f][i] = k
+
+    factors_list = []
+
+    for f, indices in factors_dict.items():
+        factors_list.append((list(f), indices))
+
+    I_neg, I_pos = _real_isolate_and_disjoin(factors_list, K, eps=eps,
+        inf=inf, sup=sup, strict=strict, basis=basis, fast=fast)
+
+    F = K.get_field()
+
+    if not zeros or not zero_indices:
+        I_zero = []
+    else:
+        if not basis:
+            I_zero = [((F.zero, F.zero), zero_indices)]
+        else:
+            I_zero = [((F.zero, F.zero), zero_indices, [K.one, K.zero])]
+
+    return sorted(I_neg + I_zero + I_pos)
+
+def _disjoint_p(M, N, strict=False):
+    """Check if Mobius transforms define disjoint intervals. """
+    a1, b1, c1, d1 = M
+    a2, b2, c2, d2 = N
+
+    a1d1, b1c1 = a1*d1, b1*c1
+    a2d2, b2c2 = a2*d2, b2*c2
+
+    if a1d1 == b1c1 and a2d2 == b2c2:
+        return True
+
+    if a1d1 > b1c1:
+        a1, c1, b1, d1 = b1, d1, a1, c1
+
+    if a2d2 > b2c2:
+        a2, c2, b2, d2 = b2, d2, a2, c2
+
+    if not strict:
+        return a2*d1 >= c2*b1 or b2*c1 <= d2*a1
+    else:
+        return a2*d1 >  c2*b1 or b2*c1 <  d2*a1
+
+def _real_isolate_and_disjoin(factors, K, eps=None, inf=None, sup=None, strict=False, basis=False, fast=False):
+    """Isolate real roots of a list of polynomials and disjoin intervals. """
+    I_pos, I_neg = [], []
+
+    for i, (f, k) in enumerate(factors):
+        for F, M in dup_inner_isolate_positive_roots(f, K,
+                eps=eps, inf=inf, sup=sup, fast=fast, mobius=True):
+            I_pos.append((F, M, k, f))
+
+        for G, N in dup_inner_isolate_negative_roots(f, K,
+                eps=eps, inf=inf, sup=sup, fast=fast, mobius=True):
+            I_neg.append((G, N, k, f))
+
+    for i, (f, M, k, F) in enumerate(I_pos):
+        for j, (g, N, m, G) in enumerate(I_pos[i+1:]):
+            while not _disjoint_p(M, N, strict=strict):
+                f, M = dup_inner_refine_real_root(f, M, K, steps=1, fast=fast, mobius=True)
+                g, N = dup_inner_refine_real_root(g, N, K, steps=1, fast=fast, mobius=True)
+
+            I_pos[i+j+1] = (g, N, m, G)
+
+        I_pos[i] = (f, M, k, F)
+
+    for i, (f, M, k, F) in enumerate(I_neg):
+        for j, (g, N, m, G) in enumerate(I_neg[i+1:]):
+            while not _disjoint_p(M, N, strict=strict):
+                f, M = dup_inner_refine_real_root(f, M, K, steps=1, fast=fast, mobius=True)
+                g, N = dup_inner_refine_real_root(g, N, K, steps=1, fast=fast, mobius=True)
+
+            I_neg[i+j+1] = (g, N, m, G)
+
+        I_neg[i] = (f, M, k, F)
+
+    if strict:
+        for i, (f, M, k, F) in enumerate(I_neg):
+            if not M[0]:
+                while not M[0]:
+                    f, M = dup_inner_refine_real_root(f, M, K, steps=1, fast=fast, mobius=True)
+
+                I_neg[i] = (f, M, k, F)
+                break
+
+        for j, (g, N, m, G) in enumerate(I_pos):
+            if not N[0]:
+                while not N[0]:
+                    g, N = dup_inner_refine_real_root(g, N, K, steps=1, fast=fast, mobius=True)
+
+                I_pos[j] = (g, N, m, G)
+                break
+
+    field = K.get_field()
+
+    I_neg = [ (_mobius_to_interval(M, field), k, f) for (_, M, k, f) in I_neg ]
+    I_pos = [ (_mobius_to_interval(M, field), k, f) for (_, M, k, f) in I_pos ]
+
+    if not basis:
+        I_neg = [ ((-v, -u), k) for ((u, v), k, _) in I_neg ]
+        I_pos = [ (( u,  v), k) for ((u, v), k, _) in I_pos ]
+    else:
+        I_neg = [ ((-v, -u), k, f) for ((u, v), k, f) in I_neg ]
+        I_pos = [ (( u,  v), k, f) for ((u, v), k, f) in I_pos ]
+
+    return I_neg, I_pos
 
 OO = 'OO' # Origin of (re, im) coordinate system
